refactor: split `utils.mjs`, rename `parseLine` to `parseCommand`, move to own utils file
